\name{correctGrowthRateMonocot}
\alias{correctGrowthRateMonocot}
\title{Correcting root growth rates}
\description{This function corrects the growth rate values stored in a data frame returned by \code{\link{architect}}.}
\usage{
correctGrowthRateMonocot(x)}
\arguments{
  \item{x}{A data frame returned by \code{\link{architect}}.}}
\value{
Returns a data frame with updated growth rate values.}
\author{Benjamin M. Delory}
\seealso{\code{\link{architect}}.}